requirements.txt
numpy==1.16.2
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow-addons==0.4.0
tensorflow-gpu==2.0.0b1
Pillow==7.0.0
